fix(jwk): adds --local=false to the DB calls
---

CI is showing 'Executing on local database' and the commands are failing. It also says 'To execute on your remote database, add a --remote flag to your wrangler command.', but that does not seem to work. However, --local=false seems to do the trick when I run it on my machine.

[skip ci]